split modules
