+++ components/infrared/impl/build.gradle.kts
@@ -21,7 +21,7 @@ dependencies {
     implementation(projects.components.bridge.pbutils)
     implementation(projects.components.bridge.synchronization.api)
 
-    implementation(projects.components.remoteControls.grid.api)
+    implementation(projects.components.remoteControls.grid.main.api)
 
     implementation(projects.components.core.di)
     implementation(projects.components.core.log)
