--- components/remote-controls/grid/api/build.gradle.kts
@@ -1,15 +1,0 @@
-plugins {
-    id("flipper.android-compose")
-}
-
-android.namespace = "com.flipperdevices.remotecontrols.grid.api"
-
-dependencies {
-    implementation(projects.components.bridge.dao.api)
-
-    implementation(projects.components.core.ui.decompose)
-    implementation(projects.components.keyedit.api)
-
-    implementation(libs.compose.ui)
-    implementation(libs.decompose)
-}
