+++ components/remote-controls/grid/create-control/api/src/main/kotlin/com/flipperdevices/remotecontrols/api/CreateControlDecomposeComponent.kt
@@ -1,0 +1,20 @@
+package com.flipperdevices.remotecontrols.api
+
+import com.arkivanov.decompose.ComponentContext
+import com.flipperdevices.bridge.dao.api.model.FlipperKeyPath
+import com.flipperdevices.keyedit.api.NotSavedFlipperKey
+import com.flipperdevices.ui.decompose.ScreenDecomposeComponent
+
+abstract class CreateControlDecomposeComponent(
+    componentContext: ComponentContext
+) : ScreenDecomposeComponent(componentContext) {
+    fun interface Factory {
+        operator fun invoke(
+            componentContext: ComponentContext,
+            savedKey: FlipperKeyPath,
+            originalKey: NotSavedFlipperKey,
+            onFinished: (FlipperKeyPath) -> Unit,
+            onFailed: () -> Unit
+        ): CreateControlDecomposeComponent
+    }
+}
