--- components/remote-controls/grid/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/grid/remote/presentation/model/GridNavigationConfig.kt
@@ -1,23 +1,0 @@
-package com.flipperdevices.remotecontrols.impl.grid.remote.presentation.model
-
-import com.flipperdevices.bridge.dao.api.model.FlipperKeyPath
-import com.flipperdevices.keyedit.api.NotSavedFlipperKey
-import kotlinx.serialization.Serializable
-
-@Serializable
-sealed class GridNavigationConfig {
-    @Serializable
-    data class Rename(val notSavedFlipperKey: NotSavedFlipperKey) : GridNavigationConfig()
-
-    @Serializable
-    data class SavedControl(val keyPath: FlipperKeyPath) : GridNavigationConfig()
-
-    @Serializable
-    data class ServerControl(val id: Long) : GridNavigationConfig()
-
-    @Serializable
-    data class Configuring(
-        val keyPath: FlipperKeyPath,
-        val notSavedFlipperKey: NotSavedFlipperKey
-    ) : GridNavigationConfig()
-}
