+++ components/remote-controls/grid/main/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/grid/main/model/GridNavigationConfig.kt
@@ -1,0 +1,23 @@
+package com.flipperdevices.remotecontrols.impl.grid.main.model
+
+import com.flipperdevices.bridge.dao.api.model.FlipperKeyPath
+import com.flipperdevices.keyedit.api.NotSavedFlipperKey
+import kotlinx.serialization.Serializable
+
+@Serializable
+sealed class GridNavigationConfig {
+    @Serializable
+    data class Rename(val notSavedFlipperKey: NotSavedFlipperKey) : GridNavigationConfig()
+
+    @Serializable
+    data class SavedControl(val keyPath: FlipperKeyPath) : GridNavigationConfig()
+
+    @Serializable
+    data class ServerControl(val id: Long) : GridNavigationConfig()
+
+    @Serializable
+    data class Configuring(
+        val keyPath: FlipperKeyPath,
+        val notSavedFlipperKey: NotSavedFlipperKey
+    ) : GridNavigationConfig()
+}
