+++ components/remote-controls/grid/main/impl/src/main/res/values/strings.xml
