+++ components/remote-controls/grid/remote/api/src/main/kotlin/com/flipperdevices/remotecontrols/grid/remote/api/RemoteGridScreenDecomposeComponent.kt
@@ -1,0 +1,20 @@
+package com.flipperdevices.remotecontrols.grid.remote.api
+
+import com.arkivanov.decompose.ComponentContext
+import com.flipperdevices.keyedit.api.NotSavedFlipperKey
+import com.flipperdevices.remotecontrols.api.model.GridControlParam
+import com.flipperdevices.ui.decompose.DecomposeOnBackParameter
+import com.flipperdevices.ui.decompose.ScreenDecomposeComponent
+
+abstract class RemoteGridScreenDecomposeComponent(
+    componentContext: ComponentContext
+) : ScreenDecomposeComponent(componentContext) {
+    fun interface Factory {
+        operator fun invoke(
+            componentContext: ComponentContext,
+            param: GridControlParam.Id,
+            onBack: DecomposeOnBackParameter,
+            onSaveKey: (NotSavedFlipperKey) -> Unit
+        ): RemoteGridScreenDecomposeComponent
+    }
+}
