+++ components/remote-controls/grid/saved/impl/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="empty_page">Page is empty!</string>
+</resources>
