+++ components/rootscreen/impl/build.gradle.kts
@@ -28,7 +28,7 @@ dependencies {
     implementation(projects.components.bridge.dao.api)
     implementation(projects.components.changelog.api)
 
-    implementation(projects.components.remoteControls.grid.api)
+    implementation(projects.components.remoteControls.grid.main.api)
     implementation(projects.components.remoteControls.main.api)
     implementation(projects.components.remoteControls.brands.api)
