+++ instances/android/app/build.gradle.kts
@@ -49,8 +49,14 @@ dependencies {
     implementation(projects.components.remoteControls.brands.impl)
     implementation(projects.components.remoteControls.categories.api)
     implementation(projects.components.remoteControls.categories.impl)
-    implementation(projects.components.remoteControls.grid.api)
-    implementation(projects.components.remoteControls.grid.impl)
+    implementation(projects.components.remoteControls.grid.createControl.api)
+    implementation(projects.components.remoteControls.grid.createControl.impl)
+    implementation(projects.components.remoteControls.grid.main.api)
+    implementation(projects.components.remoteControls.grid.main.impl)
+    implementation(projects.components.remoteControls.grid.remote.api)
+    implementation(projects.components.remoteControls.grid.remote.impl)
+    implementation(projects.components.remoteControls.grid.saved.api)
+    implementation(projects.components.remoteControls.grid.saved.impl)
     implementation(projects.components.remoteControls.main.api)
     implementation(projects.components.remoteControls.main.impl)
     implementation(projects.components.remoteControls.setup.api)
