+++ settings.gradle.kts
@@ -288,8 +288,14 @@ include(
     ":components:remote-controls:brands:api",
     ":components:remote-controls:categories:impl",
     ":components:remote-controls:categories:api",
-    ":components:remote-controls:grid:impl",
-    ":components:remote-controls:grid:api",
+    ":components:remote-controls:grid:main:impl",
+    ":components:remote-controls:grid:main:api",
+    ":components:remote-controls:grid:create-control:impl",
+    ":components:remote-controls:grid:create-control:api",
+    ":components:remote-controls:grid:remote:impl",
+    ":components:remote-controls:grid:remote:api",
+    ":components:remote-controls:grid:saved:impl",
+    ":components:remote-controls:grid:saved:api",
     ":components:remote-controls:setup:impl",
     ":components:remote-controls:setup:api",
